# [python]: Import polars-core, not polars in dependencies