--- a/py-geopolars/cargo.toml
+++ b/py-geopolars/cargo.toml
@@ -9,7 +9,7 @@ name = "geopolars"
 crate-type = ["cdylib"]
 
 [dependencies]
-polars = "0.21.1"
+polars-core = "0.21.1"
 polars-arrow = "0.21.1"
 pyo3 = { version = "0.16.4", features = ["extension-module"] }
 geopolars = { path = "../geopolars" }