I thought this would make compile times faster (since it doesn't compile `polars-time`, `polars-ops`, `polars-io`, or  `polars-lazy`), but it looks inconclusive below

```
rm -rf target/release && time maturin build --release
```

This branch:
```
4006.51s user 202.46s system 856% cpu 8:11.40 total
```

`master`:
```
3896.06s user 207.13s system 801% cpu 8:31.87 total
```